requirements.txt
altair==4.1.0
ccxt==1.89.94
GitPython==3.1.27
matplotlib==3.5.0
more_itertools==8.13.0
numpy==1.21.4
pandas==1.3.4
pandas_profiling==3.2.0
plotly==5.3.1
requests==2.26.0
Rx==3.2.0
streamlit==1.10.0
streamlit_ace==0.1.1
streamlit_pandas_profiling==0.1.3
websockets==9.1
streamlit-aggrid
Jinja2
websocket_client
requests
ccxt
python-decouple
werkzeug